fix: rename remaining folders to match naming convention
