+++ src/Components/Cart/checkout_button.tsx
@@ -9,7 +9,7 @@
 "use client";
 
 import { useState } from "react";
-import { cn } from "@/lib/utils";
+import { cn } from "@/Lib/utils";
 import { useCartStore } from "@/Store/use_cart_store";
 
 /**
